% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_asian.R
\name{am_pn_d}
\alias{am_pn_d}
\title{Calculate Risk-Neutral Probability for Down Movement}
\usage{
am_pn_d(u, d, r, n)
}
\arguments{
\item{u}{Numeric. The up factor representing the relative increase in the stock price.}

\item{d}{Numeric. The down factor representing the relative decrease in the stock price.}

\item{r}{Numeric. The risk-free rate (expressed as a factor, e.g., 1.05 for 5%).}

\item{n}{Integer. The current step in the binomial tree (\code{n > 1}).}
}
\value{
A numeric value representing the risk-neutral probability of a down movement.
}
\description{
This function calculates the risk-neutral probability of a down movement
(\eqn{p_n^d}) in a binomial tree at step \code{n}.
}
\examples{
am_pn_d(u = 1.01, d = 0.54, r = 0.90, n = 5)
}
