% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indiff_branch_probs_cpp}
\alias{indiff_branch_probs_cpp}
\title{Branch probabilities of the four-point shock scheme}
\usage{
indiff_branch_probs_cpp(rho)
}
\arguments{
\item{rho}{Correlation between the price and impact shocks, in [-1, 1].}
}
\value{
Numeric vector of length 4 holding p_(+,+), p_(+,-), p_(-,+), p_(-,-).
}
\description{
Branch probabilities of the four-point shock scheme
}
\keyword{internal}
