% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indiff_discount_helpers_cpp}
\alias{indiff_discount_helpers_cpp}
\title{Cash-discount helpers of the self-financing recursion}
\usage{
indiff_discount_helpers_cpp(r_cont, T_mat, N)
}
\arguments{
\item{r_cont}{Continuously compounded rate.}

\item{T_mat}{Maturity.}

\item{N}{Number of time steps.}
}
\value{
List with \code{dt}, \code{delta_r}, \code{beta} (length N+1) and
  \code{beta_delta} (length N).
}
\description{
Cash-discount helpers of the self-financing recursion
}
\keyword{internal}
