% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indiff_has_openmp_cpp}
\alias{indiff_has_openmp_cpp}
\title{Report whether the package was compiled with OpenMP support}
\usage{
indiff_has_openmp_cpp()
}
\value{
\code{TRUE} when OpenMP is available.
}
\description{
Report whether the package was compiled with OpenMP support
}
\keyword{internal}
