% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_asian.R
\name{print.geometric_asian_mc}
\alias{print.geometric_asian_mc}
\title{Print method for geometric_asian_mc objects}
\usage{
\method{print}{geometric_asian_mc}(x, ...)
}
\arguments{
\item{x}{A geometric_asian_mc object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for geometric_asian_mc objects
}
