% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemna_vorst_arithmetic.R
\name{print.kemna_vorst_arithmetic}
\alias{print.kemna_vorst_arithmetic}
\title{Print Method for Kemna-Vorst Arithmetic Results}
\usage{
\method{print}{kemna_vorst_arithmetic}(x, ...)
}
\arguments{
\item{x}{Object of class "kemna_vorst_arithmetic"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Kemna-Vorst Arithmetic Results
}
