% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemna_vorst_arithmetic.R
\name{summary.kemna_vorst_arithmetic}
\alias{summary.kemna_vorst_arithmetic}
\title{Summary Method for Kemna-Vorst Arithmetic Results}
\usage{
\method{summary}{kemna_vorst_arithmetic}(object, ...)
}
\arguments{
\item{object}{Object of class "kemna_vorst_arithmetic"}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for Kemna-Vorst Arithmetic Results
}
